Add check of all downloaded data, match best track to maxvort track.
Have a look at 2005 to make sure that best tracks are being correctly
matched to maxvort tracks. It is working for most best track hurricanes
over western Atlantic. However, some best tracks correspond to multiple
maxvort tracks (i.e. along their length a different maxvort track).

+++ get_data.py
@@ -69,7 +69,10 @@ def download_file(url, output_dir, local_name=None):
 	local_name = '%s/%s'%(output_dir, url.split('/')[-1])
 
     print(local_name)
-    urllib.urlretrieve(url, local_name)
+    if os.path.exists(local_name):
+        print('Already exists, skipping')
+    else:
+        urllib.urlretrieve(url, local_name)
 
 def download_c20_range(start_year, end_year):
     for year in range(start_year, end_year + 1):
